--- codes/150b6770-7f0d-4aad-b263-b973316af454.cpp
@@ -1,5 +1,0 @@
-#include <iostream>
-int main() {
-    std::cout << "Hello, World!" << std::endl;
-    return 0;
-}
