--- codes/65ed92c2-b122-47d2-b9ce-a746c17a14ab.cpp
@@ -1,5 +1,0 @@
-#include <iostream>
-int main() {
-    std::cout << "Hello, World!" << std::endl;
-    return 0;
-}
