--- codes/cpp/f4efb068-7c26-4c08-b31f-ccdf468f9400.cpp
@@ -1,5 +1,0 @@
-#include <iostream>
-int main() {
-    std::cout << "Hello, World!" << std::endl;
-    return 0;
-}
