--- generate file/cpp/codes/cpp/99365f2b-e9cf-434b-a423-bb6b0c94a2be.cpp
@@ -1,5 +1,0 @@
-#include <iostream>
-int main() {
-    std::cout << "Hello, World!" << std::endl;
-    return 0;
-}
