+++ models/Users.js
@@ -4,18 +4,32 @@ const userSchema = new mongoose.Schema({
     firstname: {
         type: String,
         default: null,
+        required: true,
     },
     lastname: {
         type: String,
         default: null,
+        required: true,
     },
     email: {
         type: String,
         unique: true,
+        required: true,
     },
     password: {
         type: String,
+        required: true,
+    },
+    role: {
+        type: String,
+        enum: ["user", "admin", "problem_creator", "contest_creator"],  // Role options
+        default: "user",  // Default role assigned at registration
+    },
+    createdAt: {
+        type: Date,
+        default: Date.now,  // Automatically set the registration date
     },
 });
 
-module.exports = mongoose.model("user", userSchema);
+const User = mongoose.model('User', userSchema);
+module.exports = User;
