--- models/problem.js
@@ -1,31 +1,0 @@
-const mongoose = require('mongoose');
-
-const problemSchema = mongoose.Schema({
-    title: {
-        type: String,
-        required: true
-    },
-    description: {
-        type: String,
-        required: true
-    },
-    difficulty: {
-        type: String,
-        enum: ['Easy', 'Medium', 'Hard'],
-        required: true
-    },
-    testCases: [{
-        input: {
-            type: String,
-            required: true
-        },
-        expectedOutput: {
-            type: String,
-            required: true
-        }
-    }]
-}, {
-    timestamps: true
-});
-
-module.exports = mongoose.model('Problem', problemSchema);
