+++ package.json
@@ -22,13 +22,15 @@
   "homepage": "https://github.com/sayed-amin/Online-Judge#readme",
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.7.1",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "nodemon": "^3.0.1",
+    "mongoose": "^8.8.3",
     "uuid": "^9.0.0"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.7"
   }
 }
