+++ utils/jwt.js
@@ -1,0 +1,7 @@
+const jwt = require('jsonwebtoken');
+
+const generateToken = (userId) => {
+  return jwt.sign({ userId }, process.env.SECRET_KEY, { expiresIn: '1h' });
+};
+
+module.exports = { generateToken };
